Documentation, separation of concerns, and more
Created documentation for app entry points (layout, provider, router)

Created documentation for floating overlay.

Created documentation for network-graph-context.

Created documentation for all hooks in the network-graph feature.

Deprecated 'network-graph-utils' and moved helper logic into hooks.
