+++ client/src/app/index.tsx
@@ -2,6 +2,10 @@ import { AppProvider} from "./provider.tsx";
 import { AppRouter } from './router.tsx';
 import './index.css';
 
+/**
+ *  Root of the app
+ */
+
 export const App = ()=> {
   return (
     <>
