+++ client/src/app/routes/landing.tsx
@@ -2,6 +2,11 @@ import { NetworkGraphProvider } from "@/features/NetworkGraph/stores/network-gra
 import { NetworkGraph } from "@/features/NetworkGraph";
 import { NetworkGraphToolbar } from "@/features/NetworkGraph/components/network-graph-toolbar.tsx";
 
+/**
+ *  Landing page component.
+ *
+ *  Holds the Network Graph feature currently.
+ */
 const Landing = () => {
     return (
         <>
