+++ client/src/components/layouts/app-layout.tsx
@@ -2,6 +2,9 @@ import { Outlet } from "react-router-dom";
 import { Header } from "../ui/header";
 import "./app-layout.css";
 
+/**
+ *  Provides current layout for the application. Renders the header above any child components passed to it.
+ */
 const AppLayout = () => {
     return (
         <div id={"appContainer"}>
