+++ client/src/features/NetworkGraph/config/paths.ts
@@ -1,0 +1,3 @@
+import { PROJECTS_DIRECTORY, TARGET_PROJECT } from "@/config/app.ts";
+
+export const BASE_PATH = `/${PROJECTS_DIRECTORY}/${TARGET_PROJECT}`;
