+++ client/src/features/NetworkGraph/hooks/useNetworkGraphData.ts
@@ -1,37 +1,47 @@
 import { useEffect, useState } from 'react';
-import { PROJECTS_DIRECTORY, TARGET_PROJECT } from "@/config/app.ts";
-import type {NetworkGraphData} from "@/types/network-graph.types.ts";
+import { BASE_PATH } from "@/features/NetworkGraph/config/paths.ts";
+
+import type { NetworkGraphData } from "@/types/network-graph.types.ts";
 
 const GRAPH_FILES = ['file-graph-data.json', 'issue-graph-data.json', 'mail-graph-data.json', 'person-graph-data.json'];
 
-export function useGraphJsonFiles(){
+type NetworkGraphJsonData = {[filename: string]: NetworkGraphData};
+
+/**
+ *  Loads all network graph data for project specified in user.config.yaml
+ */
+export function useNetworkGraphData(){
     const [graphData, setGraphData] = useState<NetworkGraphData>({ nodes: [], links: [] });
 
     useEffect(() => {
         async function load() {
-            const result = await readAllJsonsFromDirectory();
+            const jsonData: NetworkGraphJsonData = await readAllJsonsFromDirectory(BASE_PATH);
 
+            // Initialize an empty NetworkGraphData object
             const compiledGraphData : NetworkGraphData = { nodes: [], links: [] };
 
-            for (const value of Object.values(result)) {
+            // For each file, extract the nodes
+            for (const value of Object.values(jsonData)) {
                 compiledGraphData.nodes.push(... value.nodes);
                 compiledGraphData.links.push(... value.links);
             }
 
             setGraphData(compiledGraphData);
         }
+
         load();
     }, []);
 
     return graphData;
 }
 
-async function readAllJsonsFromDirectory() {
-    const BASE_PATH = `/${PROJECTS_DIRECTORY}/${TARGET_PROJECT}`;
-
+/**
+ * Loads all .json files from project specified in user.config.yaml
+ */
+async function readAllJsonsFromDirectory(filePath: string): Promise<NetworkGraphJsonData> {
     const entries = await Promise.all(
         GRAPH_FILES.map(async (fileName) => {
-            const res = await fetch(`${BASE_PATH}/${fileName}`);
+            const res = await fetch(filePath + '/' + fileName);
 
             if (!res.ok) {
                 throw new Error(`Failed to load ${fileName}: ${res.status} ${res.statusText}`);
@@ -45,3 +55,4 @@ async function readAllJsonsFromDirectory() {
     // object keyed by file name:
     return Object.fromEntries(entries);
 }
+
