+++ client/src/features/NetworkGraph/hooks/useNetworkGraphSimulation.ts
@@ -1,8 +1,7 @@
 import { useEffect, type RefObject } from "react";
 import { forceSimulation, forceLink, forceX, forceY, forceCollide, forceManyBody } from "d3";
 import type { Simulation } from "d3";
-import type { Node, Link, HubLink } from "@/types/network-graph.types.ts"
-import { nodeGroupCenters, buildHubs, buildHubLinks, clearForces } from "@/features/NetworkGraph/lib/network-graph-utils.ts";
+import type { Node, Link, HubLink, Group } from "@/types/network-graph.types.ts"
 
 interface UseNetworkGraphSimulationArgs {
     nodes: Node[];
@@ -10,6 +9,20 @@ interface UseNetworkGraphSimulationArgs {
     canvasRef: RefObject<HTMLCanvasElement | null>;
 }
 
+type NodeGroupCenters = {
+    people: [number, number];
+    mail: [number, number];
+    file: [number, number];
+    issue: [number, number];
+}
+
+/**
+ *  Runs physics simulation to determine initial position of nodes and links in the Network Graph
+ *
+ *  Note:
+ *  - D3.js force simulation is used to calculate the positions of all nodes and links
+ *  - Nodes and link objects are mutated by D3.js force simulation to contain position values x and y
+ */
 export function useNetworkGraphSimulation({ nodes, links, canvasRef }: UseNetworkGraphSimulationArgs) {
     useEffect(() => {
         if (!canvasRef.current) return;
@@ -22,7 +35,7 @@ export function useNetworkGraphSimulation({ nodes, links, canvasRef }: UseNetwor
         const nodeRadiusMultiplier = 11;
         const nodePadding = 6;
 
-        const c = nodeGroupCenters(canvas.width, canvas.height);
+        const c = calculateNodeGroupCenters(canvas.width, canvas.height);
         const hubs = buildHubs(nodes);
         const hubLinks: HubLink[] = buildHubLinks(nodes, hubs);
 
@@ -44,4 +57,47 @@ export function useNetworkGraphSimulation({ nodes, links, canvasRef }: UseNetwor
             sim.stop();
         };
     }, [nodes, links, canvasRef]);
+}
+
+function calculateNodeGroupCenters(canvasWidth: number, canvasHeight: number): NodeGroupCenters {
+    return {
+        people: [canvasWidth * 0.5, canvasHeight * 0.4],
+        mail:   [canvasWidth * 0.8, canvasHeight * 0.3],
+        file:   [canvasWidth * 0.25, canvasHeight * 0.4],
+        issue:  [canvasWidth * 0.8, canvasHeight * 0.6],
+    }
+}
+
+function buildHubs(nodes: Node[]) {
+    const hubs: Partial<Record<Group, Node>> = {};
+
+    for (const n of nodes) {
+        const existing = hubs[n.group];
+        if (!existing || n.value > existing.value) {
+            hubs[n.group] = n;
+        }
+    }
+
+    return hubs;
+}
+
+function buildHubLinks(nodes: Node[], hubs: Partial<Record<Group, Node>>) {
+    return nodes
+        .filter(n => hubs[n.group] && !isHub(n, hubs as Record<Group, Node>))
+        .map(n => ({ source: hubs[n.group]!, target: n }));
+}
+
+function isHub (d: Node, hubs: Record<Group, Node>) {
+    return hubs[d.group] === d;
+}
+
+function clearForces(sim: Simulation<Node, Link>) {
+    sim
+        .force('x', null)
+        .force('y', null)
+        .force('collide', null)
+        .force('charge', null)
+        .force('link', null)
+        .force('hubLinks', null)
+        .velocityDecay(1); // stop inertia
 }
