+++ client/src/features/NetworkGraph/stores/network-graph-context.tsx
@@ -1,6 +1,6 @@
 import { type ReactNode, createContext, useContext } from 'react';
 import { type NetworkGraphData } from '@/types/network-graph.types.ts';
-import {useGraphJsonFiles} from "@/features/NetworkGraph/hooks/useNetworkGraphJsonData.ts";
+import {useNetworkGraphData} from "@/features/NetworkGraph/hooks/useNetworkGraphData.ts";
 
 const NetworkGraphContext = createContext<NetworkGraphContextValue | null>(null);
 
@@ -8,6 +8,9 @@ interface NetworkGraphContextValue {
     data: NetworkGraphData;
 }
 
+/**
+ *  Allows a child component of the NetworkGraphProvider to use the context.
+ */
 export function useNetworkGraph() {
     const ctx = useContext(NetworkGraphContext);
 
@@ -18,10 +21,13 @@ export function useNetworkGraph() {
     return ctx;
 }
 
+/**
+ *  Provider of context for the Network Graph
+ *
+ *  Calls useGraphJsonFiles to load and provide the .json data for the Network Graph.
+ */
 export function NetworkGraphProvider({ children }: { children: ReactNode }) {
-    const data  = useGraphJsonFiles();
-
-
+    const data  = useNetworkGraphData();
 
     const value: NetworkGraphContextValue = {
         data
