+++ client/src/utils/type-guards.ts
@@ -1,0 +1,3 @@
+export function isNullOrUndefined<T>(value: T | null | undefined): value is null | undefined  {
+    return value == null;
+}
